# src/routes/static/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "UnisonHT",
        "version": "1.0.0",
        "description": "UnisonHT web server.",
        "license": {
            "name": "MIT",
            "url": "https://raw.githubusercontent.com/joeferner/unisonht/master/LICENSE.txt"
        }
    },
    "paths": {
        "/api/v1/mode": {
            "get": {
                "operationId": "getMode",
                "summary": "Get current mode",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "standard response": {
                                        "value": {
                                            "mode": "tv"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/mode/{modeName}": {
            "post": {
                "operationId": "switchMode",
                "summary": "Switch to the given mode",
                "parameters": [
                    {
                        "name": "modeName",
                        "in": "path",
                        "description": "Name of the mode to switch to",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "standard response": {
                                        "value": {
                                            "previousMode": "off",
                                            "mode": "tv"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Mode not found"
                    }
                }
            }
        },
        "/api/v1/button/{buttonName}": {
            "post": {
                "operationId": "pressButton",
                "summary": "Press the button first use the current mode then using the global buttons",
                "parameters": [
                    {
                        "name": "buttonName",
                        "in": "path",
                        "description": "Name of the button to press",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "standard response": {}
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Button not found"
                    }
                }
            }
        }
    }
}